Allow customizable redirection behavior.
Allow users to intercept and perform influence the redirection
themselves. This adds a `makeRequest` option that can be used to
do just that. For now, my plan is to leave it undocumented,
because I want to get a release out, and don't want to commit
to this API decision just yet. Specifically, I want make some
progress on #19, and #12 before committing to exactly how
it will behave.