Fix hanging application when the final response is destroyed prematurely
This calls `destroy()` on the redirected responses
as soon as they are not needed anymore.

This fixes the case where the application would hang if `destroy()`
would be called on the final response without it being consumed
completely.

This also slightly improves performance,
as the redirected responses can be cleaned up immediately.